separate List and NavList
