+++ src/components/NavList/NavList.less
@@ -1,0 +1,35 @@
+@import (reference) 'List/List';
+
+.nav-list {
+
+  .list;
+
+  cursor: pointer;
+
+  .current {
+    position: relative;
+  }
+
+  .current::before {
+    background-color:   var(--blue-4);
+    inset-block-end:    calc(-0.5 * var(--border-width));
+    content:            '';
+    inset-inline-start: 0;
+    position:           absolute;
+    inset-block-start:  0;
+    width:              0.25em;
+  }
+
+  .current:first-child::before {
+    border-start-start-radius: var(--border-radius);
+  }
+
+  .current:last-child::before {
+    border-end-start-radius: var(--border-radius);
+  }
+
+  li:hover {
+    background-color: var(--gray-1);
+  }
+
+}
